refactor: extract prompt rendering to prompt_renderer.py
Move _load_global_instructions, _parse_agent_hooks, _build_required_steps,
_load_review_section, _load_continuation_section,
_load_intervention_context_for_prompt, and _render_prompt (~230 lines) out of
scheduler.py into a dedicated prompt_renderer.py module.

scheduler.py re-imports and re-exports all moved names so external callers
(tests, jobs.py) are unaffected.
